+++ src/OpenTelemetry/Metrics/CircularBufferBuckets.cs
@@ -205,6 +205,8 @@ internal sealed class CircularBufferBuckets
                 Consolidate(array, (offset + (uint)(index - begin)) % capacity, (offset + (uint)((index >> 1) - (begin >> 1))) % capacity);
             }
 
+            // Don't merge below call into above for loop.
+            // Merging causes above loop to be infinite if end = int.MaxValue, because index <= int.MaxValue is always true.
             Consolidate(array, (offset + (uint)(end - begin)) % capacity, (offset + (uint)((end >> 1) - (begin >> 1))) % capacity);
         }