+++ src/OpenTelemetry.Instrumentation.Http/CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## Unreleased
 
+## 1.0.0-rc9.6
+
+Released 2022-Aug-18
+
+* Updated to use Activity native support from `System.Diagnostics.DiagnosticSource`
+  to set activity status.
+  ([#3118](https://github.com/open-telemetry/opentelemetry-dotnet/issues/3118))
+  ([#3555](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3555))
+
 * Changed activity source name from `OpenTelemetry.HttpWebRequest`
   to `OpenTelemetry.Instrumentation.Http.HttpWebRequest` for `HttpWebRequest`s
   and from `OpenTelemetry.Instrumentation.Http`