+++ src/OpenTelemetry/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+## 1.4.0-alpha.2
+
+Released 2022-Aug-18
+
 * Added `Sdk.CreateLoggerProviderBuilder` method and support for dependency
   injection scenarios when configuring `OpenTelemetryLoggerProvider`
   ([#3504](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3504))